REF: Split InterMol-based GROMACS parser out into new module
